+++ clang-tools-extra/clang-tidy/misc/MySecondCheckCheck.cpp
@@ -1,0 +1,40 @@
+//===--- MySecondCheckCheck.cpp - clang-tidy ------------------------------===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+
+#include "MySecondCheckCheck.h"
+#include "clang/AST/ASTContext.h"
+#include "clang/ASTMatchers/ASTMatchFinder.h"
+
+using namespace clang::ast_matchers;
+
+namespace clang {
+namespace tidy {
+namespace misc {
+
+void MySecondCheckCheck::registerMatchers(MatchFinder *Finder) {
+  if(!getLangOpts().CPlusPlus) {
+    return;
+  }
+
+  Finder->addMatcher(varDecl(allOf(hasType(asString("struct B")), hasDescendant(cxxNewExpr()))).bind("id"), this);
+}
+
+void MySecondCheckCheck::check(const MatchFinder::MatchResult &Result) {
+  // FIXME: Add callback implementation.
+  const auto *MatchedDecl = Result.Nodes.getNodeAs<VarDecl>("id");
+  if (MatchedDecl->getName().startswith("awesome_"))
+    return;
+  diag(MatchedDecl->getLocation(), "function %0 is insufficiently awesome")
+      << MatchedDecl
+      << FixItHint::CreateInsertion(MatchedDecl->getLocation(), "awesome_");
+}
+
+} // namespace misc
+} // namespace tidy
+} // namespace clang